# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "QuestEd API",
    "version": "1.4.1",
    "description": "Comprehensive API specification for QuestEd educational platform",
    "contact": {
      "name": "QuestEd Development Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.quested.example.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean"},
          "data": {"type": "object"},
          "message": {"type": "string"},
          "errors": {"type": "object"},
          "timestamp": {"type": "string", "format": "date-time"}
        }
      },
...
    }
  },
  "paths": {
...
    "/api/units": {
      "get": {
        "summary": "List curriculum units",
        "description": "Retrieve available curriculum units with optional filtering",
        "parameters": [
          {
            "name": "school_id",
            "in": "query",
            "schema": {"type": "integer"},
            "description": "Filter by school ID"
          },
          {
            "name": "subject_id",
            "in": "query",
            "schema": {"type": "integer"},
            "description": "Filter by subject ID"
          },
          {
            "name": "difficulty_level",
            "in": "query",
            "schema": {"type": "integer", "minimum": 1, "maximum": 5},
            "description": "Filter by difficulty level"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {"type": "integer", "default": 1},
            "description": "Page number for pagination"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {"type": "integer", "default": 20},
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Units retrieved successfully",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ApiResponse"}
              }
            }
          }
        }
      }
    },
    "/api/units/select": {
      "post": {
        "summary": "Select unit for learning",
        "description": "Student selects a unit for learning in a specific class",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["unit_id", "class_id"],
                "properties": {
                  "unit_id": {"type": "integer"},
                  "class_id": {"type": "integer"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unit selected successfully",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ApiResponse"}
              }
            }
          }
        }
      }
    },
...
  }
}